+++ src/converters/video.ts
@@ -2,7 +2,15 @@ import { DomNode } from 'html-parse-stringify';
 import {
   DocumentBodyVideo,
   DocumentBodyVideoBlock,
+  DocumentBodyVideoProperties,
 } from '../models/blocks/document-body-video.js';
+import { DocumentElementLength } from '../models/blocks/document-element-length.js';
+import {
+  getHeightWithUnit,
+  getStyleKeyValues,
+  getWidthWithUnit,
+} from './table-properties.js';
+import { StyleAttribute } from '../models/html/style-attribute.js';
 
 export const generateVideoBlock = (
   iframeElement: DomNode,
@@ -14,7 +22,60 @@ export const generateVideoBlock = (
 };
 
 const generateVideo = (iframeElement: DomNode): DocumentBodyVideo => {
-  return {
-    url: iframeElement.attrs?.src || '',
-  };
+  const videoUrl = iframeElement.attrs?.src || '';
+
+  const properties = getVideoProperties(iframeElement);
+
+  return Object.assign({}, { url: videoUrl }, properties && { properties });
+};
+
+const getVideoProperties = (
+  iframeElement: DomNode,
+): DocumentBodyVideoProperties | undefined => {
+  let videoProperties: DocumentBodyVideoProperties | undefined;
+
+  if (iframeElement.attrs) {
+    let width: DocumentElementLength | undefined;
+    let height: DocumentElementLength | undefined;
+
+    if (iframeElement.attrs?.width) {
+      width = getWidthWithUnit({ width: iframeElement.attrs?.width });
+    }
+
+    if (iframeElement.attrs?.height) {
+      height = getHeightWithUnit({ height: iframeElement.attrs?.height });
+    }
+
+    if (iframeElement.attrs?.style) {
+      const styleKeyValues = getStyleKeyValues(iframeElement);
+
+      if (
+        Object.prototype.hasOwnProperty.call(
+          styleKeyValues,
+          StyleAttribute.Width,
+        )
+      ) {
+        width = getWidthWithUnit(styleKeyValues);
+      }
+
+      if (
+        Object.prototype.hasOwnProperty.call(
+          styleKeyValues,
+          StyleAttribute.Height,
+        )
+      ) {
+        height = getHeightWithUnit(styleKeyValues);
+      }
+    }
+
+    if (width) {
+      videoProperties = Object.assign(
+        {},
+        width && { width },
+        height && { height },
+      );
+    }
+  }
+
+  return videoProperties;
 };