+++ src/models/blocks/document-body-video.ts
@@ -1,3 +1,5 @@
+import { DocumentElementLength } from './document-element-length.js';
+
 export interface DocumentBodyVideoBlock {
   type: 'Video';
   video: DocumentBodyVideo;
@@ -5,4 +7,9 @@ export interface DocumentBodyVideoBlock {
 
 export interface DocumentBodyVideo {
   url: string;
+  properties?: DocumentBodyVideoProperties;
+}
+
+export interface DocumentBodyVideoProperties {
+  width?: DocumentElementLength;
 }