+++ test/convert-html-to-blocks/convert-html-to-blocks.spec.ts
@@ -20,6 +20,16 @@ describe('convert-html-to-blocks', function () {
     it('image', test); // knowledge-administration-ui / image-block-utils.spec.ts / should generate image block from image html
     it('image-with-properties', test); // knowledge-administration-ui / image-block-utils.spec.ts / should generate image block with properties from image html
     it('image-with-embedded-src', test);
+    it('image-with-properties-max-alt-length', async function (this: Context, options: HtmlConverterOptions = {}) {
+      const dirName = join('test', ...this.test!.titlePath());
+      const html = (await readFile(join(dirName, 'input.html')))
+        .toString()
+        .replace(/\r/gi, '');
+
+      const actualJson = convertHtmlToBlocks(html, options);
+
+      expect(actualJson?.[0]?.image?.properties?.altText?.length).to.equal(200);
+    });
   });
 
   describe('list', function () {
@@ -92,6 +102,10 @@ describe('convert-html-to-blocks', function () {
 
   describe('video', function () {
     it('simple', test); // knowledge-administration-ui / video-block-utils.spec.ts / should generate video blocks from html
+    it('video-with-properties', test); // knowledge-administration-ui / video-block-utils.spec.ts / should generate video blocks from html with properties
+    it('video-with-properties-em-width-from-style', test); // knowledge-administration-ui / video-block-utils.spec.ts / should generate video blocks from html with properties
+    it('video-with-properties-px-width-from-style', test); // knowledge-administration-ui / video-block-utils.spec.ts / should generate video blocks from html with properties
+    it('video-with-properties-percentage-width-from-style', test); // knowledge-administration-ui / video-block-utils.spec.ts / should generate video blocks from html with properties
   });
 
   /*
@@ -120,7 +134,15 @@ describe('convert-html-to-blocks', function () {
     it('handleWidthWithUnits-disabled', testHandleWidthWithUnitDisabled);
     // handleWidthWithUnits - for images
     it('handleWidthWithUnits-percentage-image', testHandleWidthWithUnitEnabled);
+    it(
+      'handleWidthWithUnits-percentage-image-from-width-attribute',
+      testHandleWidthWithUnitEnabled,
+    );
     it('handleWidthWithUnits-pixel-image', testHandleWidthWithUnitEnabled);
+    it(
+      'handleWidthWithUnits-pixel-image-from-width-attribute',
+      testHandleWidthWithUnitEnabled,
+    );
     it(
       'handleWidthWithUnits-pixel-image-no-input-unit',
       testHandleWidthWithUnitEnabled,